Split models module into separate files
