+++ .flake8
@@ -16,6 +16,7 @@ per-file-ignores =
     orchest-sdk/python/orchest/__init__.py: F401
     orchest-cli/orchestcli/cli.py: W505, E501
     services/orchest-api/app/app/core/sessions/__init__.py: F401
+    services/orchest-api/app/app/models/__init__.py: F401
     services/orchest-ctl/app/orchest/__init__.py: F401
 
 # black is set to 88
